[TEST] AudioMux with VocalSorcery Mics
NOT WORKING!

Mic Array on Tile0 Header J14
Contains left over changes/experiments before switching to
lib_mic_array vanilla demo.

I believe this code had valid signaling to/from the mics, but it
was not clear if the audio was being sampled correctly.

The USB audio was working when using a test waveform (sinusoid),
but that was about it.